backend/requirements.txt
django>=5.0
djangorestframework>=3.15
django-cors-headers>=4.3
python-dotenv>=1.0
requests>=2.31
numpy>=1.26
scipy>=1.12
